Move AddVolumeAndVerify() to a util directory that can be imported by other tests
