+++ tests/BUILD.bazel
@@ -233,6 +233,7 @@ go_test(
         "//tests/testsuite:go_default_library",
         "//tests/usb:go_default_library",
         "//tests/util:go_default_library",
+        "//tests/util/storage-util:go_default_library",
         "//tests/virtctl:go_default_library",
         "//tests/virtiofs:go_default_library",
         "//tests/watcher:go_default_library",
